MP3 Backend support
The MP3 strategy using a format string to determine its output.
The strategy works and unit tests have been created.

An error in the NameExtensionHelper has been fixed.  If a file does
not have an extension then the extension field was set to the
whole filename.  This has been fixed.