AsyncLogic Project committed
This project is testing out the async capabilities of C#.
The idea behind this logic is that we treat 'true' as a
terminating program and 'false' as a looping program.
We can 'and' and 'or' these programs as a form of logic.
Note that 'not' cannot be expressed due to the Halting Problem.

Basic syntax expressions and some tests have been added.

The visitor pattern has been used extensively.