AsyncLogic backend changes
BoolValue now supports false values.  So instead of looping, if we
know that something is false we just return it.  This increases
performance, but it did required restructing the behaviour for 'or'.

Before, as soon as one of 'or's operands finished we could return it.
Now we have to check that it is not false.

The reason we did not use false values before was because we could
not demonstrate looping.  Now that we have 'exists' and 'the' we
can loop properly, so these performance enhancements can now be
added.

Unit tests checking the correctness of 'and' and 'or' under true,
false, and looping has been added.