AsyncLogic pair support
Added support for pairs in this logic.

The evaluator has some difficulties with pairs and projections,
so have added a new value type which is a potential pair.
Its elements are tasks, which need to be run in order to get the
true value.  However, these may explode.

New methods have been added to values to fully evaluate them.
Note that they return tasks, so using a timeout may be necessary.

Unit tests have been added for pairs.