AsyncLogic Changed the handling of pairs
The difficulty with projections of pairs is that they can be of
any type.  Before we had them deriving from Expression.  Now an
extra class of each type has been constructed.  Interfaces for left
and right projections have been added so that the visitors do not
have to handle all the cases separately.