AsyncLogic New Expression Type
Added a new expression type to help parse expressions.  Also includes
a new parser to parse text into this type.

Added a new class for types, plus a parser.

Parser for sequents and class also added.

Working on a visitor to determine in a parsed expression is type
correct.

Note: when parsing with this new parser bound variables must be
given a type.  This means we can avoid type inference, at the
expense of convenience.