AsyncLogic ToExpressionVisitor Added
A new visitor will convert from parsed expressions to our original
expressions.  Also finished the type visitor.

There is a lot of complexity involved with having two different
types of expressions.  The original type, while somewhat more
typesafe, is more awkward to use.  In the future it may be refactored.

A very tiny change to the documentation for FileRenamer.