AsyncLogic Fixing failure unit test
The previous version of the ParserFailure unit test would pass
if only one of the tests raised an exception.  This has been
changed so that all the tests must raise an exception.

A csv file of the tests has been added to the project.

Some changes to the parser have been made:
T -> True
F -> False
The order of <,> and <=, >= have been changed so they parse correctly.

The parser still has some problems, such as 6 == The ... does not
work, but 6 == (The ... )  does work.